CW - Peacemaker's Blade (#129)
* wrote a simple version of the card

it was a copy and paste from pacifism, but it not correct yet. need to change the canPlay to see if the character is attacking

* need to look at the context object to see if this info is in there
* need to write test for the card

* first draft of the card

I realize that canPlay was the wrong method I wanted canAttach, this gave
me the info about the character and not just the context object. needs tests
for:

* cant attach to an attacking character
* can attach to defender
* character with it cant be declared as attacker
* can be declared as defender
* can be moved in as an attacker

* wrote the test and fixed copy paste error

change the effect of the card to `cardCannot(declareAttacker)` instead of
partipiate as attacker, and wrote test for:
* declaring the character as an attacker
* moving the character in
* delaring as defender
* attaching to an attacker

test for being declared is failing, probably the effect isnt getting passed
down correctly

* fixed declaring bug

typo in the effect I wrote declareAttacker instead of declareAsattacker
after corrected all test pass