CW - Ramparts of stone (#140)
* first draft of card class

it should target who ever is the attacking player, then prompt them to
either bow participating character or discard card from hand

* renamed and fixed typo

I wrote the class name and file name as rampants when it should have been
ramparts

* got the first test work

found a bug in the bow character option. I need to redetermine the correct
player (attacking one). first test just makes sense it gives the attacking
player the choices. next needs test for each options

* basic test for the two options

added test to makes that that bow option only bows character in the conflict. and another test that the select cards are discarded

* added test for FoJ and alway targeting the attacker

added two test:
1. makes sure finger of jade doesnt cancel the card
2. makes sure that it is always the attacking player even if it is the cards
controller

* change how it got the characters to bow

because of a request on github I was told that their is a function on the
conflict to get a players partipating character. so this is now used instead
of looking at all the card that player has and filitering for characters in
the conflict

* fixed lint error

after last commit I dont use the CardTypes enum so I had to remove the import