CW - Hallowed ground (#166)
* first draft of the fire ring effect

this is the implementation if the opponent has the fire ring. I need to write
test. the test will need to make sure that the player cant place fate on
conflict character or dynasty charaters(in case they have the fire ring in
the dynasty phase)

* fire ring effect

added test to see if the card works on conflict characters but it isnt working

* add air ring persistence effect and tests

the air ring is based on breach of etiquette, two basic test were added to
make sure it is working correctly, but it isnt

* update with the code review from zarzuckett

test for the air ring are still failing. I fixed the bug with the fire ring
even though the player only has one choice you still need to click the zero
prompt. added test for multiple conflicts and righteous magistrate

* fixed typo in test

* show that shadowland hunterworks with it