RaW - Our For Does Not Wait (#224)
* had to delete the branch

there were merge conflict from rally keyword that I didnt know how to get rid
of. so this is the card plus two simple tests

* added more test

added tests for:
* checking it has eight cards to choose from
* the card is placed in the correct province
* and the card stacks in the province

* forgot to shuffle the deck

* added test for max one per conflict

I missed that the card could only be used once per conflict, this was pointed
out in a code review and a test was added to make sure it is working
correctly

* using deck search

the card uses deck search instead a handler, the deck search action class
had to be changed to make sure the card is faceup in the province. the
were changed to make sure the character is faceup