RaW - Garanto Guardian (#229)
* first version of the card

this is mostly a copy paste from the new mitsu. just has a reaction and allow
the player to select only rings that are in the attacked provinces elements
array

* added test

so I added some basic test for the card as well as the toshi ranbo tests

* test for if the ring is claimed

in the pass other card have had a lot bug report for not being able to select
the ring if it is claimed typically by the opponent. so here is a test to make
sure that bug does not happen for this card